requirement.txt
awesome-slugify==1.6.5
Django==1.10.1
django-password-reset==1.0
Markdown==2.6.11
Pillow==5.0.0
redis==2.10.6
regex==2018.2.21
sorl-thumbnail==12.4.1
Unidecode==0.4.21
